libxbps: detection of orphaned packages is now 86% faster.
Instead of building deptree, then checking its all components, list of packages
to remove (`array`) is used as queue for BFS graph traversing. Speedup comes
from the fact, that if some package is still needed, all its sub-deptree can't
be removed too, and processing that sub-tree is futile.
The other change is early exit from revdeps loop. When one revdep of X package
is not orphan, it is clear that X is not orpan too.
Changes in test does only update order in which packages are selected for
removal.